refactor: consolidate time formatting functions (DRY)
- Enhanced formatTicksAsTime in common.ts to handle days, hours, and minutes
- Added TICK_DURATION_SECONDS constant to common.ts
- Removed duplicate formatTicksDuration from growth.ts
- Re-exported formatTicksAsTime as formatTicksDuration for backwards compatibility
- Added deprecation notice to formatTicksDuration alias
- Added comprehensive tests for formatTicksAsTime in time.test.ts
